Remove custom encryption class and add AuthenticateStaffUser method
This commit removes the custom sha1 encyption method in favor of the built-in CommonCrypto library found in Xcode 10. This greatly simplifies hashing of the signature string and uses a well-established and trusted library. It may have been fun implementing the cutom class, but it wasn't necessary in this case.

This commit also adds the ability to authenticate a staff user, which will be used to access protected API methods. All credentials for the Polaris API are stored in a plist file called PolarisAPI-Info that is added to a project along with the compiled framework.