Rewrite BasicDataGetResponse as Patron
Rewrite the getBasicData response object as Patron. Each value is decoded manually to remove a level of encoding (Patron.name versus Patron.basicData.name) and to use types that are not traditionally supported by the Codable protocol such as Date.

The endpoint for getBasicData was also modified so that an array of addresses is returned with the results.