Start implementation of updatePatron
Start the implementation of the updatePatron method. As written the method works but is not ideal as it overwrites some values with junk values. The method needs to be expanded to ensure that good data is not overwritten. The ability to update an address is not available as the Polaris API does not allow for a direct replacement of the address and instead requires staff intervention.