Refactor Patron struct
Refactor the Patron struct so that the style more closely resembles the refactored AuthenticatedStaffUser. These changes led to the removal of the PatronDecoder struct and the PolarisUtility struct.

The changes to Patron and the removal of PolarisUtility led to a few minor changes in other structs but nothing that changes functionality, with the exception of UpdatePatronRequest, which needs significant revisions and was not fully implemented prior to these changes.