Add PolarisAPI structure with to-do items
Create a new PolarisAPI structure with all of the API calls added as to-do items. As more API calls were being added to the Polaris struct, some of the request or response objects were getting similar names and I wanted to be able to more clearly identify an API call from a response or request object.