Fix: decoding fails
Fix some issues that caused decoding to fail.

Two of the values returned by the API are strings rather than integers, despite the content of the text being an integer value.

Two of the values returned by the API are integers rather than integers, despite only returning a value of 1 or 0 and treated as boolean values.
Three of the values returned by the API can be null values, so the properties have been changed to optional.