Begin comprehensive rewrite - Core + Bib
This begins a comprehensive rewrite that will also increase the number of PAPI endpoints supported by this package. This first update includes some of the core methods for handling network requests as well as the Bib endpoints. Existing applications should continue to use tag 0.1 until this rewrite is complete as things are bound to break along the way.